refactor: split MoqKnownSymbolsTests into partial class files by subject
The test file exceeded the 1000-line limit. Split into 6 partial class
files grouped by Moq concept:

- MoqKnownSymbolsTests.cs: infrastructure, constructors, helpers
- MoqKnownSymbolsTests.MockAndRepository.cs: Mock, MockBehavior, It, Times
- MoqKnownSymbolsTests.ReturnsAndThrows.cs: IReturns, IThrows, IReturnsResult
- MoqKnownSymbolsTests.CallbackAndSetup.cs: ICallback, ISetup, ISetupGetter
- MoqKnownSymbolsTests.RaisesAndEvents.cs: IRaiseable, IRaise
- MoqKnownSymbolsTests.Caching.cs: caching, thread-safety, BCL types

No test logic changed. All 138 MoqKnownSymbolsTests pass.
